format code and divide stimulation folder into sources and targets
