+++ components/definition/stimulationSources/NetPyNEStimulationSource.js
@@ -1,16 +1,10 @@
 import React, { Component } from 'react';
-import SelectField from 'material-ui/SelectField';
 import MenuItem from 'material-ui/MenuItem';
 import TextField from 'material-ui/TextField';
-import Tooltip from 'material-ui/internal/Tooltip';
-import FlatButton from 'material-ui/FlatButton';
-import Toggle from 'material-ui/Toggle';
-import IconMenu from 'material-ui/IconMenu';
-import RaisedButton from 'material-ui/RaisedButton';
-import clone from 'lodash.clone';
+import SelectField from 'material-ui/SelectField';
 import Utils from '../../../Utils';
-import NetPyNEField from '../../general/NetPyNEField';
 import ListComponent from '../../general/List';
+import NetPyNEField from '../../general/NetPyNEField';
 
 var PythonControlledCapability = require('../../../../../js/communication/geppettoJupyter/PythonControlledCapability');
 var PythonControlledTextField = PythonControlledCapability.createPythonControlledControl(TextField);
@@ -24,7 +18,6 @@ export default class NetPyNEStimulationSource extends React.Component {
       currentName: props.name,
       sourceType: null
     };
-    
     this.stimSourceTypeOptions = [
       { type: 'IClamp' }, 
       { type: 'VClamp' }, 
@@ -33,7 +26,7 @@ export default class NetPyNEStimulationSource extends React.Component {
       { type: 'AlphaSynapse'}
     ];
     this.handleStimSourceTypeChange = this.handleStimSourceTypeChange.bind(this);
-  }
+  };
 
   componentWillReceiveProps(nextProps) {
     if (this.state.currentName!=nextProps.name) {
@@ -47,17 +40,15 @@ export default class NetPyNEStimulationSource extends React.Component {
     var newValue = event.target.value;
     this.setState({ currentName: newValue });
     this.triggerUpdate(function () {
-      // Rename the population in Python
       Utils.renameKey('netParams.stimSourceParams', storedValue, newValue, (response, newValue) => { that.renaming = false });
       that.renaming = true;
     });
   };
   
   triggerUpdate(updateMethod) {
-    //common strategy when triggering processing of a value change, delay it, every time there is a change we reset
     if (this.updateTimer != undefined) {
       clearTimeout(this.updateTimer);
-    }
+    };
     this.updateTimer = setTimeout(updateMethod, 1000);
   };
   
@@ -66,9 +57,9 @@ export default class NetPyNEStimulationSource extends React.Component {
   };
 
   componentDidUpdate(prevProps, prevState) {
-      if (this.state.currentName != prevState.currentName) {
-          this.updateLayout();
-      };
+    if (this.state.currentName != prevState.currentName) {
+      this.updateLayout();
+    };
   };
 
   updateLayout() {
@@ -96,9 +87,10 @@ export default class NetPyNEStimulationSource extends React.Component {
           id={"sourceName"}
         />
         <br/>
+        
         <NetPyNEField id="netParams.stimSourceParams.type" className={"netpyneFieldNoWidth"} noStyle>
           <SelectField 
-            floatingLabelText="Stim type"
+            floatingLabelText="stimulation type"
             value={this.state.sourceType}
             onChange={this.handleStimSourceTypeChange}
           >
@@ -109,158 +101,188 @@ export default class NetPyNEStimulationSource extends React.Component {
             }
           </SelectField>
         </NetPyNEField>  
-      </div>);
+      </div>
+    );
+      
     if (this.state.sourceType=='IClamp'){
-      var variableContent = 
+      var variableContent = (
         <div>
           <NetPyNEField id="netParams.stimSourceParams.del">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['del']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.dur" className="listStyle">
             <PythonControlledListComponent
               model={"netParams.stimSourceParams['" + this.props.name + "']['dur']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.amp" className="listStyle">
             <PythonControlledListComponent
               model={"netParams.stimSourceParams['" + this.props.name + "']['amp']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.i">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['i']"}
             />
           </NetPyNEField>
         </div>
+      );
     } else if (this.state.sourceType=='VClamp'){
-      var variableContent = 
+      var variableContent = (
         <div>
           <NetPyNEField id="netParams.stimSourceParams.tau1">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['tau1']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.tau2">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['tau2']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.dur" className="listStyle">
             <PythonControlledListComponent
               model={"netParams.stimSourceParams['" + this.props.name + "']['dur']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.amp" className="listStyle">
             <PythonControlledListComponent
               model={"netParams.stimSourceParams['" + this.props.name + "']['amp']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.gain">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['gain']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.rstim">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['rstim']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.i">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['i']"}
             />
           </NetPyNEField>
         </div>
+      );
     } else if (this.state.sourceType=='AlphaSynapse'){
-      var variableContent = 
+      var variableContent = (
         <div>
           <NetPyNEField id="netParams.stimSourceParams.onset">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['onset']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.tau">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['tau']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.gmax">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['gmax']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.e">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['e']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.i">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['i']"}
             />
           </NetPyNEField>
         </div>
+      );
     }else if (this.state.sourceType=='NetStim'){
-      var variableContent = 
+      var variableContent = (
         <div>
           <NetPyNEField id="netParams.stimSourceParams.interval">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['interval']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.number">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['number']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.start">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['start']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.noise">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['noise']"}
             />
           </NetPyNEField>
         </div>
+      );
     } else if (this.state.sourceType=='SEClamp'){
-      var variableContent = 
+      var variableContent = (
         <div>
           <NetPyNEField id="netParams.stimSourceParams.rs">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['rs']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.dur" className="listStyle">
             <PythonControlledListComponent
               model={"netParams.stimSourceParams['" + this.props.name + "']['dur']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.amp" className="listStyle">
             <PythonControlledListComponent
               model={"netParams.stimSourceParams['" + this.props.name + "']['amp']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.i">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['i']"}
             />
           </NetPyNEField>
+          
           <NetPyNEField id="netParams.stimSourceParams.vc">
             <PythonControlledTextField
               model={"netParams.stimSourceParams['" + this.props.name + "']['vc']"}
             />
           </NetPyNEField>
         </div>
-    } 
+      ) 
+    } else {
+      var variableContent = <div/>
+    };
+    
     return (
       <div>
         {content}
         {variableContent}
       </div>
     );
-  }
-}
+  };
+};
