+++ components/definition/stimulationSources/NetPyNEStimulationThumbnail.js
@@ -1,26 +1,26 @@
 import React, { Component } from 'react';
 import FloatingActionButton from 'material-ui/FloatingActionButton';
 
-export default class NetPyNECellRuleThumbnail extends React.Component {
+export default class NetPyNEStimulationThumbnail extends React.Component {
 
   constructor(props) {
     super(props);
     this.handleClick = this.handleClick.bind(this);
-  }
+  };
 
   handleClick() {
     if (this.props.handleClick) {
       this.props.handleClick(this.props.name);
-    }
-  }
+    };
+  };
 
   render() {
-    return (
-
+    var content = (
       <FloatingActionButton className={"actionButton " + (this.props.selected ? "selectedActionButton" : "")} onClick={this.handleClick}>
         {this.props.name}
       </FloatingActionButton>
-
     );
-  }
-}
+    
+    return content;
+  };
+};
