+++ components/definition/stimulationTargets/NetPyNEStimulationTargets.js
@@ -1,59 +1,41 @@
 import React, { Component } from 'react';
-import Card, { CardHeader, CardText } from 'material-ui/Card';
-import FlatButton from 'material-ui/FlatButton';
-import RaisedButton from 'material-ui/RaisedButton';
-import FloatingActionButton from 'material-ui/FloatingActionButton';
-import FontIcon from 'material-ui/FontIcon';
-import IconButton from 'material-ui/IconButton';
 import IconMenu from 'material-ui/IconMenu';
-import MenuItem from 'material-ui/MenuItem';
-
-import NetPyNEStimulationThumbnail from './NetPyNEStimulationThumbnail';
-import NetPyNEStimulationTarget from './NetPyNEStimulationTarget';
-import NetPyNENewStimulationTarget from './NetPyNENewStimulationTarget';
-
+import Card, { CardHeader, CardText } from 'material-ui/Card';
 import Utils from '../../../Utils';
+import NetPyNEStimulationTarget from './NetPyNEStimulationTarget';
+import NetPyNEStimulationThumbnail from '../stimulationSources/NetPyNEStimulationThumbnail';
+import NetPyNENewStimulationTarget from './NetPyNENewStimulationTarget';
 
 export default class NetPyNEStimulationTargets extends React.Component {
 
   constructor(props) {
     super(props);
     this.state = {
-      drawerOpen: false,
       selectedStimulationTarget: undefined,
-      page: "main"
     };
-
     this.selectStimulationTarget = this.selectStimulationTarget.bind(this);
     this.handleNewStimulationTarget = this.handleNewStimulationTarget.bind(this);
-    
-  }
+  };
 
   selectStimulationTarget(StimulationTarget) {
     this.setState({ selectedStimulationTarget: StimulationTarget });
-  }
+  };
 
   handleNewStimulationTarget(defaultStimulationTargets) {
-    // Get Key and Value
+    console.log('NEW STIM TARGET')
     var key = Object.keys(defaultStimulationTargets)[0];
     var value = defaultStimulationTargets[key];
     var model = this.state.value;
-
-    // Get New Available ID
     var StimulationTargetId = Utils.getAvailableKey(model, key);
-
-    // Create Cell Rule Client side
     Utils.execPythonCommand('netpyne_geppetto.netParams.stimTargetParams["' + StimulationTargetId + '"] = ' + JSON.stringify(value));
-
-    // Update state
     this.setState({
       value: model,
       selectedStimulationTarget: StimulationTargetId
     });
-  }
-
+  };
 
   hasSelectedStimulationTargetBeenRenamed(prevState, currentState) {
+    console.log('HAS CHANGED')
     var currentModel = prevState.value;
     var model = currentState.value;
     //deal with rename
@@ -67,72 +49,66 @@ export default class NetPyNEStimulationTargets extends React.Component {
             if (prevState.selectedStimulationTarget != undefined) {
               if (oldP[i] == prevState.selectedStimulationTarget) {
                 return newP[i];
-              }
-            }
-          }
-        }
-      }
-    }
+              };
+            };
+          };
+        };
+      };
+    };
     return false;
-  }
-
+  };
  
   componentDidUpdate(prevProps, prevState) {
-    //we need to check if any of the three entities have been renamed and if that's the case change the state for the selection variable
+    console.log('DID UPDATED')
     var newStimulationTargetName = this.hasSelectedStimulationTargetBeenRenamed(prevState, this.state);
     if (newStimulationTargetName) {
       this.setState({ selectedStimulationTarget: newStimulationTargetName });
-    }
-  }
+    };
+  };
 
   shouldComponentUpdate(nextProps, nextState) {
+    console.log('SHOULD UPDATE')
     var itemRenamed = this.hasSelectedStimulationTargetBeenRenamed(this.state, nextState) != false;
     var newItemCreated = false;
     var selectionChanged = this.state.selectedStimulationTarget != nextState.selectedStimulationTarget;
-    var pageChanged = this.state.page != nextState.page;
-    var newModel = this.state.value == undefined;
-    if (!newModel) {
+    if (this.state.value!=undefined) {
       newItemCreated = Object.keys(this.state.value).length != Object.keys(nextState.value).length;
-    }
-    return newModel || newItemCreated || itemRenamed || selectionChanged || pageChanged;
-  }
+    };
+    return newItemCreated || itemRenamed || selectionChanged;
+  };
 
   render() {
-
     var model = this.state.value;
-    var content;
-    if (this.state.page == 'main') {
+    var StimulationTargets = [];
+    for (var c in model) {
+      StimulationTargets.push(<NetPyNEStimulationThumbnail name={c} key={c} selected={c == this.state.selectedStimulationTarget} handleClick={this.selectStimulationTarget} />);
+    };
+    var selectedStimulationTarget = undefined;
+    if (this.state.selectedStimulationTarget && Object.keys(model).indexOf(this.state.selectedStimulationTarget)>-1) {
+      selectedStimulationTarget = <NetPyNEStimulationTarget name={this.state.selectedStimulationTarget}/>;
+    };
 
-      var StimulationTargets = [];
-      for (var c in model) {
-        StimulationTargets.push(<NetPyNEStimulationThumbnail name={c} key={c} selected={c == this.state.selectedStimulationTarget} handleClick={this.selectStimulationTarget} />);
-      }
-      var selectedStimulationTarget = undefined;
-      if (this.state.selectedStimulationTarget && Object.keys(model).indexOf(this.state.selectedStimulationTarget)>-1) {
-        selectedStimulationTarget = <NetPyNEStimulationTarget name={this.state.selectedStimulationTarget}/>;
-      }
-
-      content = (
-        <CardText className={"tabContainer"} expandable={true}>
-          <div className={"details"}>
-            {selectedStimulationTarget}
+    var content = (
+      <CardText className={"tabContainer"} expandable={true}>
+        <div className={"details"}>
+          {selectedStimulationTarget}
+        </div>
+        <div className={"thumbnails"}>
+          <div className="breadcrumb">
+            <IconMenu style={{ float: 'left', marginTop: "12px", marginLeft: "18px" }}
+              iconButtonElement={
+                <NetPyNENewStimulationTarget name={selectedStimulationTarget} handleClick={this.handleNewStimulationTarget} />
+              }
+              anchorOrigin={{ horizontal: 'left', vertical: 'top' }}
+              targetOrigin={{ horizontal: 'left', vertical: 'top' }}
+            >
+            </IconMenu>
           </div>
-          <div className={"thumbnails"}>
-            <div className="breadcrumb">
-              <IconMenu style={{ float: 'left', marginTop: "12px", marginLeft: "18px" }}
-                iconButtonElement={
-                  <NetPyNENewStimulationTarget name={selectedStimulationTarget} handleClick={this.handleNewStimulationTarget} />
-                }
-                anchorOrigin={{ horizontal: 'left', vertical: 'top' }}
-                targetOrigin={{ horizontal: 'left', vertical: 'top' }}
-              >
-              </IconMenu>
-            </div>
-            <div style={{ clear: "both" }}></div>
-            {StimulationTargets}
-          </div>
-        </CardText>);
-    }
+          <div style={{ clear: "both" }}></div>
+          {StimulationTargets}
+        </div>
+      </CardText>
+    );
 
     return (
       <Card style={{ clear: 'both' }}>
@@ -143,6 +119,7 @@ export default class NetPyNEStimulationTargets extends React.Component {
           showExpandableButton={true}
         />
         {content}
-      </Card>);
-  }
-}
+      </Card>
+    );
+  };
+};
