+++ components/definition/synapses/NetPyNESynapseThumbnail.js
@@ -11,16 +11,16 @@ export default class NetPyNESynapseThumbnail extends React.Component {
   handleClick() {
     if (this.props.handleClick) {
       this.props.handleClick(this.props.name);
-    }
-  }
+    };
+  };
 
   render() {
-    return (
-
+    var content = (
       <FloatingActionButton className={"actionButton " + (this.props.selected ? "selectedActionButton" : "")} onClick={this.handleClick}>
         {this.props.name}
       </FloatingActionButton>
-
     );
-  }
-}
+    
+    return content;
+  };
+};
